fix!: change EstimatedHours and ActualHours from int to float64
The Backlog API accepts and returns decimal values (e.g. 2.5h) for
estimatedHours and actualHours. Using int caused JSON decode errors
when the API returned a fractional number.

BREAKING CHANGE: Issue.EstimatedHours and Issue.ActualHours fields
changed from int to float64. WithEstimatedHours and WithActualHours
option methods now accept float64 instead of int.

+++ internal/model/issue.go
@@ -20,8 +20,8 @@ type Issue struct {
 	Milestone      []*Version     `json:"milestone,omitempty"`
 	StartDate      string         `json:"startDate,omitempty"`
 	DueDate        string         `json:"dueDate,omitempty"`
-	EstimatedHours int            `json:"estimatedHours,omitempty"`
-	ActualHours    int            `json:"actualHours,omitempty"`
+	EstimatedHours float64        `json:"estimatedHours,omitempty"`
+	ActualHours    float64        `json:"actualHours,omitempty"`
 	ParentIssueID  int            `json:"parentIssueId,omitempty"`
 	CreatedUser    *User          `json:"createdUser,omitempty"`
 	Created        time.Time      `json:"created,omitempty"`
